+++ src/main/java/miyucomics/hexical/casting/patterns/akashic/OpClearAkashicShelf.kt
@@ -1,0 +1,29 @@
+package miyucomics.hexical.casting.patterns.akashic
+
+import at.petrak.hexcasting.api.spell.*
+import at.petrak.hexcasting.api.spell.casting.CastingContext
+import at.petrak.hexcasting.api.spell.iota.Iota
+import at.petrak.hexcasting.api.spell.mishaps.MishapBadBlock
+import at.petrak.hexcasting.common.blocks.akashic.BlockEntityAkashicBookshelf
+import at.petrak.hexcasting.common.lib.HexBlocks
+import net.minecraft.util.math.BlockPos
+
+class OpClearAkashicShelf : SpellAction {
+	override val argc = 1
+	override fun execute(args: List<Iota>, ctx: CastingContext): Triple<RenderedSpell, Int, List<ParticleSpray>> {
+		val position = args.getBlockPos(0, argc)
+		ctx.assertVecInRange(position)
+		val block = ctx.world.getBlockState(position)
+		if (!block.isOf(HexBlocks.AKASHIC_BOOKSHELF))
+			throw MishapBadBlock.of(position, "akashic_bookshelf")
+		return Triple(Spell(position), 0, listOf())
+	}
+
+	private data class Spell(val position: BlockPos) : RenderedSpell {
+		override fun cast(ctx: CastingContext) {
+			val shelf = ctx.world.getBlockEntity(position)!! as BlockEntityAkashicBookshelf
+			shelf.clearIota()
+			shelf.markDirty()
+		}
+	}
+}